Revert "Inventory restrucutring"
This reverts commit 9f56d5a469378d1ed2c255a78d33e8d93b7faafe.
