+++ FATEL_Frontend/src/app/inventory/inventory.component.css
@@ -19,49 +19,3 @@
 .content-container {
   display: flex;
 }
-
-.inventory {
-  flex: 1;
-  min-width: 70%;
-}
-
-.panel-description {
-  flex: 1;
-}
-
-table {
-  width: 100%;
-}
-
-.ta-left {
-  text-align: left !important;
-}
-
-.action-cell {
-  padding: 0;
-  height: 48px;
-  border-bottom: 1px solid rgba(0, 0, 0, .12);
-  justify-content: center;
-}
-
-.edit-action {
-  width: 100px;
-}
-
-.edit-button {
-  padding: 0;
-  margin-right: 24px;
-}
-
-.edit-button:hover {
-  background-color: var(--primary);
-  color: var(--white);
-}
-
-.action-container {
-  position: sticky;
-  flex: 1;
-  align-self: flex-start;
-  top: 0;
-  max-width: 30%;
-}
